requirements-pip.txt
imutils==0.5.3
keras==2.3.1
matplotlib==3.2.1
numpy==1.18.4
omegaconf==2.0.6
opencv-python==4.2.0.34
pillow==7.1.2
scikit-learn==0.23.1
scipy==1.4.1
tensorflow-gpu==2.2.0
